requirement.txt
pytest-playwright
pytest-asyncio
openpyxl
pillow